+++ EduFeedbackHub/CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
+## [0.1.1] - 2025-07-03
+### Added
+- Added functionality to import QS rankings data into the database
+- Enabled retrieving and displaying QS rankings from the database
+
 ## [0.1.0] - 2025-07-03
 ### Added
 - Implemented QS rankings query for three years
 - Set up initial Django project structure
+